Second and final batch
