+++ Controller.Chromecast/requirements.txt
@@ -5,3 +5,4 @@ requests
 gevent
 python-kasa
 pychromecast
+python-icc
