+++ Controller.Chromecast/src/utils.py
@@ -2,7 +2,7 @@ import os
 from flask import Flask
 import pychromecast
 from chromecastplayer import ChromecastPlayer
-from objectid import PydanticObjectId
+from icc.models import PydanticObjectId
 from repository import DeviceRepository
