+++ Controller.KasaLedStrip/src/__main__.py
@@ -1,14 +1,13 @@
 import asyncio
 from threading import Thread
-from objectid import PydanticObjectId
-from utils import initialize_bulbs, start_background_loop
-from repository import DeviceRepository, StateRepository, AnalyticsRepository
+from utils import initialize_led_strips, start_background_loop
+from repository import AnalyticsRepository, DeviceRepository, StateRepository
 from flask import Flask, Response, jsonify, request
 from flask_cors import CORS
 from kasa import SmartDeviceException
-from lightingrequest import LightingRequest
-from bulb import Bulb
+from ledstrip import LedStrip
 from gevent.pywsgi import WSGIServer
+from icc.models import PydanticObjectId, LightingRequest
 
 
 """ Flask and Repository Setup """
@@ -21,14 +20,16 @@ state_repository: StateRepository = StateRepository(app)
 analytics_repository: AnalyticsRepository = AnalyticsRepository(app)
 
 
-""" Bulbs and Asyncio Event Loop """
+""" Led Strips and Asyncio Event Loop """
 
 loop: asyncio.AbstractEventLoop = asyncio.new_event_loop()
-global bulbs
-bulbs: dict[PydanticObjectId, Bulb] = initialize_bulbs(device_repository, loop)
+global led_strips
+led_strips: dict[PydanticObjectId, LedStrip] = initialize_led_strips(
+    device_repository, loop)
 
 
 """ Error Handler """
+
 
 @app.errorhandler(404)
 def resource_not_found(e):
@@ -37,6 +38,7 @@ def resource_not_found(e):
 
 """ Health """
 
+
 @app.route("/health")
 def index() -> Response:
     return "Healthy", 200
@@ -44,10 +46,11 @@ def index() -> Response:
 
 """ Bulb Requests"""
 
+
 @app.route("/update", methods=["PUT"])
-def update_bulbs() -> Response:
-    global bulbs
-    bulbs = initialize_bulbs(device_repository, loop)
+def update_led_strips() -> Response:
+    global led_strips
+    led_strips = initialize_led_strips(device_repository, loop)
     return "Success", 200
 
 
@@ -55,18 +58,17 @@ def update_bulbs() -> Response:
 def lighting_request() -> Response:
     try:
         lighting_request = LightingRequest(**request.get_json())
-        bulb_controller = bulbs[lighting_request.target]
-        bulb_controller.set_request(lighting_request)
+        led_strip_controller = led_strips[lighting_request.target]
+        led_strip_controller.set_request(lighting_request)
+
         asyncio.run_coroutine_threadsafe(
-            bulb_controller.operation_callback_by_name[lighting_request.operation](
+            led_strip_controller.operation_callback_by_name[lighting_request.operation](
             ), loop
         )
 
-        state: bool = None
+        state: bool = False
         if lighting_request.operation != "off":
             state = True
-        else:
-            state = False
 
         state_repository.update(lighting_request.target, state)
         analytics_repository.save(lighting_request.target, state)
@@ -78,9 +80,9 @@ def lighting_request() -> Response:
 
 
 if __name__ == "__main__":
-    bulb_thread = Thread(target=start_background_loop,
-                         args=(loop,), daemon=True)
-    bulb_thread.start()
+    strip_thread = Thread(target=start_background_loop,
+                          args=(loop,), daemon=True)
+    strip_thread.start()
 
     http_server = WSGIServer(("", 8000), app)
     http_server.serve_forever()
