+++ Controller.KasaLedStrip/src/utils.py
@@ -1,5 +1,5 @@
 import asyncio
-from objectid import PydanticObjectId
+from icc.models import PydanticObjectId
 from ledstrip import LedStrip
 from repository import DeviceRepository
 
@@ -15,6 +15,7 @@ def initialize_led_strips(device_repository: DeviceRepository, loop: asyncio.Abs
         led_strips[device.id] = bulb
 
     return led_strips
+
 
 def start_background_loop(loop: asyncio.AbstractEventLoop):
     asyncio.set_event_loop(loop)
