--- Controller.KasaPlug/src/powerrequest.py
@@ -1,8 +1,0 @@
-from typing import Optional
-from pydantic import BaseModel, Field
-from objectid import PydanticObjectId
-
-
-class PowerRequest(BaseModel):
-    target: PydanticObjectId
-    operation: str
