+++ Service.Device/requirements.txt
@@ -6,3 +6,4 @@ flask-cors
 requests
 gevent
 pymongo
+python-icc
