+++ Service.Power/requirements.txt
@@ -4,6 +4,6 @@ aioflask
 requests
 gevent
 fastapi
-
 flask-pymongo
 pymongo
+python-icc
